requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
amqp==5.2.0
    # via kombu
asgiref==3.7.2
    # via django
astroid==2.15.8
    # via flake8-django
async-timeout==4.0.3
    # via redis
billiard==4.2.0
    # via celery
celery==5.3.5
    # via -r requirements.in
click==8.1.7
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.0
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.3.0
    # via celery
django==4.2.7
    # via
    #   -r requirements.in
    #   django-bootstrap5
    #   django-ckeditor
    #   django-debug-toolbar
    #   django-js-asset
    #   django-model-utils
    #   django-phonenumber-field
django-bootstrap5==23.3
    # via -r requirements.in
django-ckeditor==6.7.0
    # via -r requirements.in
django-debug-toolbar==4.2.0
    # via -r requirements.in
django-js-asset==2.2.0
    # via django-ckeditor
django-model-utils==4.3.1
    # via -r requirements.in
django-phonenumber-field==7.2.0
    # via -r requirements.in
faker==20.1.0
    # via -r requirements.in
flake8==6.1.0
    # via
    #   -r requirements.in
    #   flake8-django
    #   pep8-naming
flake8-django==1.4
    # via -r requirements.in
isort==5.12.0
    # via -r requirements.in
kombu==5.3.4
    # via celery
lazy-object-proxy==1.10.0
    # via astroid
mccabe==0.7.0
    # via flake8
pep8-naming==0.13.3
    # via -r requirements.in
phonenumbers==8.13.25
    # via -r requirements.in
pillow==10.1.0
    # via -r requirements.in
prompt-toolkit==3.0.43
    # via click-repl
pycodestyle==2.11.1
    # via flake8
pyflakes==3.1.0
    # via flake8
python-dateutil==2.8.2
    # via
    #   celery
    #   faker
redis==5.0.1
    # via -r requirements.in
six==1.16.0
    # via
    #   -r requirements.in
    #   python-dateutil
sqlparse==0.4.4
    # via
    #   django
    #   django-debug-toolbar
typing-extensions==4.9.0
    # via
    #   asgiref
    #   astroid
tzdata==2023.3
    # via celery
vine==5.1.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.12
    # via prompt-toolkit
wrapt==1.16.0
    # via astroid
